+++ core/src/api/mod.rs
@@ -1,4 +1,6 @@
 pub mod configuration;
+pub mod diagnostics;
 pub mod server;
+pub mod types;
 mod v1;
 pub mod v2;