+++ core/src/api/server.rs
@@ -8,22 +8,28 @@
 //! * `/v1/confidence/{block_number}` - returns calculated confidence for a given block number
 //! * `/v1/appdata/{block_number}` - returns decoded extrinsic data for configured app_id and given block number
 
-use crate::api::v2;
-use crate::data::Database;
-use crate::network::p2p;
-use crate::shutdown::Controller;
-use crate::types::IdentityConfig;
 use crate::{
-	api::v1,
-	network::rpc::{self},
+	api::{
+		diagnostics,
+		types::{Error, ErrorCode, InternalServerError},
+		v1, v2,
+	},
+	data::Database,
+	network::{p2p, rpc},
+	shutdown::Controller,
+	types::IdentityConfig,
 };
+
 use color_eyre::eyre::WrapErr;
 use futures::{Future, FutureExt};
+use hyper::StatusCode;
 use std::{net::SocketAddr, str::FromStr};
+use tracing::error;
 use tracing::info;
-use warp::{Filter, Reply};
+use warp::{Filter, Rejection, Reply};
 
 use super::configuration::{APIConfig, SharedConfig};
+use super::types::WsClients;
 
 pub struct Server<T: Database> {
 	pub db: T,
@@ -31,7 +37,7 @@ pub struct Server<T: Database> {
 	pub identity_cfg: IdentityConfig,
 	pub version: String,
 	pub node_client: rpc::Client<T>,
-	pub ws_clients: v2::types::WsClients,
+	pub ws_clients: WsClients,
 	pub shutdown: Controller<String>,
 	pub p2p_client: p2p::Client,
 }
@@ -56,15 +62,19 @@ impl<T: Database + Clone + Send + Sync + 'static> Server<T> {
 			self.node_client.clone(),
 			self.ws_clients.clone(),
 			self.db.clone(),
-			self.p2p_client.clone(),
 		);
+		let diagnostics_api = diagnostics::routes(self.p2p_client);
 
 		let cors = warp::cors()
 			.allow_any_origin()
 			.allow_header("content-type")
 			.allow_methods(vec!["GET", "POST", "DELETE"]);
 
-		let routes = health_route().or(v1_api).or(v2_api).with(cors);
+		let routes = health_route()
+			.or(v1_api)
+			.or(v2_api)
+			.or(diagnostics_api)
+			.with(cors);
 
 		let addr = SocketAddr::from_str(format!("{host}:{port}").as_str())
 			.wrap_err("Unable to parse host address from config")
@@ -76,4 +86,24 @@ impl<T: Database + Clone + Send + Sync + 'static> Server<T> {
 
 		server
 	}
+}
+
+pub async fn handle_rejection(error: Rejection) -> Result<impl Reply, Rejection> {
+	if error.find::<InternalServerError>().is_some() {
+		return Ok(StatusCode::INTERNAL_SERVER_ERROR.into_response());
+	}
+	Err(error)
+}
+
+pub fn log_internal_server_error(result: Result<impl Reply, Error>) -> Result<impl Reply, Error> {
+	if let Err(Error {
+		error_code: ErrorCode::InternalServerError,
+		cause: Some(error),
+		message,
+		..
+	}) = result.as_ref()
+	{
+		error!("{message}: {error:#}");
+	}
+	result
 }