+++ core/src/api/v2/ws.rs
@@ -1,11 +1,11 @@
-use super::{
-	transactions,
-	types::{Payload, Request, Response, Status, Version, WsClients, WsError, WsResponse},
-};
+use super::transactions;
 use crate::{
 	api::{
 		configuration::SharedConfig,
-		v2::types::{Error, Sender},
+		types::{
+			Error, Payload, Request, Response, Sender, Status, Version, WsClients, WsError,
+			WsResponse,
+		},
 	},
 	data::{Database, RpcNodeKey},
 	utils::spawn_in_span,